Refactor: 프로젝트 구조 변경
- 프로젝트 구조 변경
- Config 클래스 제거
