+++ app/src/main/java/com/survivalcoding/stopwatch/PrefsController.kt
@@ -3,12 +3,11 @@ package com.survivalcoding.stopwatch
 import android.content.Context
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import com.survivalcoding.stopwatch.Config.Companion.KEY_LOG_ARRAY_LIST
 import java.lang.reflect.Type
 
 class PrefsController(private val context: Context) {
     fun putLogArrayList(logList: List<Long>) {
-        val prefs = context.getSharedPreferences(Config.KEY_PREFS, Context.MODE_PRIVATE)
+        val prefs = context.getSharedPreferences(KEY_PREFS, Context.MODE_PRIVATE)
         val editor = prefs.edit()
         val json = Gson().toJson(logList)
         editor.putString(KEY_LOG_ARRAY_LIST, json)
@@ -16,7 +15,7 @@ class PrefsController(private val context: Context) {
     }
 
     fun restoreLogArrayList(): List<Long> {
-        val prefs = context.getSharedPreferences(Config.KEY_PREFS, Context.MODE_PRIVATE)
+        val prefs = context.getSharedPreferences(KEY_PREFS, Context.MODE_PRIVATE)
         val json = prefs.getString(KEY_LOG_ARRAY_LIST, null)
         return if (json == null) {
             listOf()
